requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0.6
djangorestframework==3.15.1
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==4.0.1
google==3.0.0
google-auth==2.29.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kubernetes==29.0.0
MarkupSafe==2.1.5
mysql==0.0.3
mysql-connector-python==8.4.0
mysqlclient==2.2.4
oauthlib==3.2.2
protobuf==5.27.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pylogrus==0.4.0
pymongo==4.7.3
python-bootstrap==0.0.5
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==69.5.1
shortuuid==1.0.13
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.43.0
